pyproject.toml
[tool.poetry]
name = "llm-rags"
version = "0.1.0"
description = "LLM + RAG experiments and evaluation framework"
authors = ["Omid Maghsoudi <omidh.usa@gmail.com>"]
readme = "README.md"
packages = [
    { include = "common_modules", from = "src" },
    { include = "RAGs", from = "src" },
    { include = "eval_models", from = "src" },
    { include = "LLMs_training", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# Core ML / Data
numpy = "<2"
torch = ">=2.3.0,<2.4.0"
torchvision = ">=0.17.2"
pandas = ">=2.2.1,<3.0.0"

# NLP / Evaluation
datasets = ">=4.4.2,<5.0.0"
nltk = ">=3.9.2,<4.0.0"
sacrebleu = ">=2.5.1,<3.0.0"
rouge = ">=1.0.1,<2.0.0"
bert-score = ">=0.3.13,<0.4.0"
sentence-transformers = ">=2.7,<3.0"

# Vector DB / Search
faiss-cpu = "*"
chromadb = ">=1.3.7,<2.0.0"

# LLM / RAG
ollama = ">=0.6.1,<0.7.0"
huggingface-hub = ">=0.34.0,<1.0"

# LangChain
langchain = ">=1.0.0,<2.0.0"
langchain-core = ">=1.2.2,<2.0.0"
langchain-community = ">=0.3.31,<0.4.0"
langchain-openai = ">=1.1.6,<2.0.0"
langchain-huggingface = ">=1.2.0,<2.0.0"
langchain-text-splitters = "==1.1.0"

# Document processing
unstructured = { version = ">=0.18.21,<0.19.0", extras = ["pdf"] }
bs4 = ">=0.0.2,<0.0.3"

# Config / Utils
httpx = ">=0.27,<0.28"
pyyaml = ">=6.0.1,<7.0.0"
omegaconf = ">=2.3.0,<3.0.0"
hydra-core = ">=1.3.2,<2.0.0"
load-dotenv = ">=0.1.0,<0.2.0"

# Audio / Speech
gtts = ">=2.5.4,<3.0.0"
faster-whisper = ">=1.2.1,<2.0.0"
diffusers = ">=0.36.0,<0.37.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
